notifications table - show tooltip on click
